time: port to Gaia
For monotonic time, Gaia provides processes with the CPU tick rate in
the auxillary vector, and the rdtsc instruction is sufficient to
determine the current time (and highly efficient!). For realtime clocks,
we reach out to the clock service.

time::sleep is currently not great on Gaia, and will be rewritten when a
better service is available for this purpose.

Signed-off-by: Drew DeVault <sir@cmpwn.com>