all: improve some strerror docs
When there's a single error type which aliases void, the docs reflect
that it always returns the same string. In the future, many more
functions will return nomem, and this likely won't be included in the
module's error type, so it's good to be specific in the docs.

A typo in glob::strerror's docs is also fixed.