hare::module: Remove unused recursive flags
There is a fundamental problem when submodules are gathered recursively,
but their dependencies are not. If a module was collected as a "mere"
dependency, but turns out to be part of the submodules tree or searched
in a subsequent call to gather(), it will eventually be promoted from
"is_dep" to "!is_dep".

The problem is that at this point it was inserted before its potential
dependencies in the reverse-topologically-sorted slice of modules. To
avoid the complications of a solution in the hare::module code we can
instead move the needle the call site, which was already achieved with
'hare deps -D' managing its non-recursive outputs.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>