os::exec: fix memory leak with free_strlist
free_strlist didn't free the last member of the slice, under the
assumption that it will always be null. But this may not necessarily be
true if a strlist is being constructed and one of the appends nomems. In
this case, the most recently appended string wouldn't be freed.

This is fixed by just iterating over the entire slice. `free` is a no-op
for null pointers, so there's no need for any special handling here.

Signed-off-by: Sebastian <sebastian@sebsite.pw>