debug: Use .skip instead of .balign for altstack definition
balign did something weird to the binary layout which caused valgrind
to fail symbol resolution.

The .balign directive also has somewhat unexpected behaviour when used
like this -- it can allocate up to twice the requested size because it's
trying to align the end of the allocation to a 16K boundary rather than
allocating 16K exactly.

This change does not affect the size of Hare binaries.

Signed-off-by: Alexey Yerin <yyp@disroot.org>