Add support for DragonFly BSD
All tests pass:

    make check HARETEST_INCLUDE="slow integration"

    598 passed; 0 failed; 598 completed in 61.711468923s

The porting process went rather smooth, thanks to the fact that harec
(and QBE) are written in portable C so there was no need cross-compile
the bootstrap compiler from another operating system. Also, it's
very helpful to have test cases in place. Apart from that, Hare is quite
pleasant to work with.

It would be great to get DragonFly support on SourceHut, so that
future changes won't break DragonFly support.

This work is based on the platform-specific code for FreeBSD (x86_64).
DragonFly started out as a fork from FreeBSD 20+ years ago and their
similarities are still visible - 95% of the stdlib hare code is the
same.  BUT there are subtile differences:

* Several structs have a slightly different layout. "stat" for instance.

* A few defines have different values or are missing in DragonFly.

* A few syscalls have an additional argument (which is 0).  This is
  probably due to historical reasons when switching from IA32 to x86_64
  architecture and due to alignment. This applies to lseek(), ftruncate()
  and mmap() and possibly some others.

* For the same reason the mmap() syscall has to pass one
  argument on the stack to the kernel. This is hand-coded in assembly
  instead of exposing a syscall7() as it's the only syscall that
  needs 7 arguments.

* There is no sysctlbyname() in DragonFly. We have to emulate it via
  sysctlnametomib() + sysctl().

* DragonFly does not come out of the box with a /etc/mime.types.
  There is a port/package "mime-support" which installs it
  under /usr/local/etc/mime.types.

* DragonFly does not (yet) ship with a
  /usr/share/zoneinfo/leap-seconds.list. I am trying to get this into
  the next version of DragonFly, as without that file some tests of
  time::chrono will fail.  The file can be obtained from NetBSD [1].

In order to bootstrap hare, you also need my DragonFly related patches
for harec.

[1]: https://github.com/NetBSD/src/blob/43d53c97b01cfa07cfc6fe2818c7ecafb7fcee66/external/public-domain/tz/dist/leap-seconds.list

Signed-off-by: Michael Neumann <mneumann@ntecs.de>