Added time::sleep_until, added optional clock argument to time::sleep. sleep_until is implemented using clock_nanosleep(...TIMER_ABSTIME...).
Openbsd doesn't have clock_nanosleep(), so time::sleep_until has been
backfilled with a version that uses nanosleep() instead. Beware, though,
that this version has a flaw. If, for example, the system is suspended
between the now() and sleep() calls, the whole suspend time will be added
to the total time spent sleeping. If clock_nanosleep() is implemented in
openbsd, this will be fixed.

v6: fixed a very silly error in the backfilled openbsd sleep_until().

Signed-off-by: Eric Toombs <storehouse@toombs.earth>