hare::module::gather: dedupe on names
Rather than paths. Fixes referencing the same module from two different
parts of $HAREPATH in one build:

./
├── a/
│   └── b/
│       ├── b.ha:
│       │ export let x = 0;
│       └
├── main.ha:
│ use ab = a::b;
│ use b;
│
│ export fn main() void = {
│       ab::x = 42;
│       assert(b::x == 0);
│ };
│
├── run.sh+
│ #!/bin/sh -eu
│ eval "$(hare version -vv)"
│ env HAREPATH="$PWD/a:$HAREPATH" hare run "$@"
└

Note that a::b and b:: are linked in separately, since b::x's symbol
needs to be 'b.x' and a::b::x needs to be 'a.b.x'. It's a bit weird, but
that's the only semantics that makes sense, and I don't think it's a
good idea to disallow this

Fixes: https://todo.sr.ht/~sircmpwn/hare/910

Signed-off-by: Ember Sawady <ecs@d2evs.net>